Refactor context hooks into separate files and clean up exports
