+++ src/components/points/PointSettingsModal.tsx
@@ -10,7 +10,7 @@ import {
   Zap,
 } from 'lucide-react';
 import { useCallback, useEffect, useState } from 'react';
-import { useAuth } from '../../contexts/AuthContext';
+import { useAuth } from '../../hooks/useAuth';
 import { useGroup, useUserGroups } from '../../hooks/useGroup';
 import { PointRule, pointsService } from '../../lib/points';
 import { LoadingSpinner } from '../common/LoadingSpinner';
