+++ src/components/task/MentionInput.tsx
@@ -4,7 +4,7 @@ import { Button } from '@/components/ui/button';
 import { Textarea } from '@/components/ui/textarea';
 import { Badge } from '@/components/ui/badge';
 import { cn } from '@/components/ui/utils';
-import { useAuth } from '@/contexts/AuthContext';
+import { useAuth } from '@/hooks/useAuth';
 import { useData } from '@/contexts/DataContext';
 import { useGroup } from '@/hooks/useGroup';
 import { GroupMember } from '@/types/group';
