+++ src/components/ui/badge.tsx
@@ -14,4 +14,3 @@ function Badge({ className, variant, ...props }: BadgeProps) {
 }
 
 export { Badge };
-export default Badge;
