+++ src/hooks/useTasks.ts
@@ -1,6 +1,6 @@
 import { Timestamp } from 'firebase/firestore';
 import { useCallback, useEffect, useState } from 'react';
-import { useAuth } from '../contexts/AuthContext';
+import { useAuth } from '../hooks/useAuth';
 import { taskService } from '../lib/firestore';
 import { pointsService } from '../lib/points';
 import {
