+++ src/pages/StatisticsEnhanced.tsx
@@ -32,7 +32,7 @@ import { StatisticsInsights } from '../components/statistics/StatisticsInsights'
 import { GlassCard } from '../components/ui/GlassCard';
 import { WaveButton } from '../components/ui/WaveButton';
 import { Typography } from '../components/ui/typography';
-import { useAuth } from '../contexts/AuthContext';
+import { useAuth } from '../hooks/useAuth';
 import { useGroup, useUserGroups } from '../hooks/useGroup';
 import { useTasks } from '../hooks/useTasks';
 import { PointStats, pointsService } from '../lib/points';
